Layout & Variable Updates
‣ Renamed `%nav-hover` to `%hover-transition` as this is more accurately
describes what the placeholder does, and it is not necessarily specific
to the navigation.
‣ Renamed `hover` mixin to `underline-effect`. For the same reasons as
above, trying to keep everything as semantic as possible.
‣ Renamed `%margin-bottom` to `%section separator`. Again, this is to
keep things semantic.
‣ Slightly adjust the footer's bottom margin on larger screen sizes. We
reduce the bottom margin by 5 pixels to compensate for the margin included in
the underline-effect mixin.
‣ Adjust max-width of contact-links to 50% as they extended the entire width of
the parent container, making them too easy to accidentally click.
‣ Add bottom border to the skills section header by default. This just
gives a clear separation from the rest of the content in the section.